Vanilla Javascript way to make sticky element
It uses another element in the dom flow to use as offset for the scroll element.

The element will stick to botom right of window until that offset limit element is reached. Then it will stay put in that place.

on scroll up, it will become sticky again as it crosses the limit threshold.